리펙토링 : Answers 일급 객체 추가하여 로직 위임
